backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via pytest
beautifulsoup4==4.11.1
    # via
    #   -r requirements.in
    #   django-livereload-server
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   pytest-celery
certifi==2023.7.22
    # via
    #   elastic-transport
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   nltk
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==39.0.1
    # via
    #   pyjwt
    #   pyopenssl
    #   pysaml2
defusedxml==0.7.1
    # via
    #   djangosaml2
    #   pysaml2
    #   python3-openid
dj-rest-auth[with_social]==2.2.7
    # via -r requirements.in
django==4.1.10
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-livereload-server
    #   django-revproxy
    #   djangorestframework
    #   djangosaml2
django-allauth==0.52.0
    # via dj-rest-auth
django-livereload-server==0.4
    # via -r requirements.in
django-revproxy @ git+https://github.com/jazzband/django-revproxy.git@1defbb2dad5c0632391d54bcd3dbdaeabf46266a
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
djangosaml2==1.5.6
    # via -r requirements.in
elastic-transport==8.4.0
    # via elasticsearch
elasticsearch==8.6.0
    # via -r requirements.in
elementpath==4.1.1
    # via xmlschema
et-xmlfile==1.1.0
    # via openpyxl
execnet==1.9.0
    # via pytest-xdist
fst-pso==1.8.1
    # via pyfume
fuzzytm==2.0.5
    # via gensim
gensim==4.3.0
    # via -r requirements.in
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
kombu==5.2.4
    # via celery
langcodes==3.3.0
    # via -r requirements.in
language-data==1.1
    # via -r requirements.in
lxml==4.9.1
    # via -r requirements.in
marisa-trie==0.7.8
    # via language-data
miniful==0.0.6
    # via fst-pso
nltk==3.8.1
    # via -r requirements.in
numpy==1.24.1
    # via
    #   -r requirements.in
    #   fst-pso
    #   fuzzytm
    #   gensim
    #   pandas
    #   pyfume
    #   scikit-learn
    #   scipy
    #   simpful
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.2
    # via -r requirements.in
packaging==23.0
    # via pytest
pandas==1.5.3
    # via
    #   -r requirements.in
    #   fuzzytm
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.36
    # via click-repl
psycopg2==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyfume==0.2.25
    # via fuzzytm
pyjwt[crypto]==2.6.0
    # via django-allauth
pyopenssl==23.1.1
    # via pysaml2
pypdf2==3.0.1
    # via -r requirements.in
pysaml2==7.3.1
    # via djangosaml2
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-xdist
pytest-celery==0.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-xdist==3.1.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   pandas
    #   pysaml2
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7
    # via
    #   celery
    #   djangorestframework
    #   pandas
    #   pysaml2
redis==4.4.4
    # via -r requirements.in
regex==2022.10.31
    # via nltk
requests==2.31.0
    # via
    #   django-allauth
    #   pysaml2
    #   requests-oauthlib
    #   simpful
requests-oauthlib==1.3.1
    # via django-allauth
scikit-learn==1.2.1
    # via -r requirements.in
scipy==1.10.0
    # via
    #   fuzzytm
    #   gensim
    #   pyfume
    #   scikit-learn
    #   simpful
simpful==2.9.0
    # via pyfume
six==1.16.0
    # via
    #   click-repl
    #   django-livereload-server
    #   python-dateutil
smart-open==6.3.0
    # via gensim
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
textdistance==4.5.0
    # via -r requirements.in
threadpoolctl==3.1.0
    # via scikit-learn
tornado==6.3.2
    # via django-livereload-server
tqdm==4.64.1
    # via
    #   -r requirements.in
    #   nltk
urllib3==1.26.13
    # via
    #   django-revproxy
    #   elastic-transport
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
xmlschema==2.2.3
    # via pysaml2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
